Removing useless code
